MXFS 0.89.88 — warning-free build; four latent bugs the warnings hid are fixed
- P99-IGET probes declared mxfs_inode_disk_di_size() with two parameters against a three-parameter definition that writes through the third; with instrumentation on they wrote through a garbage pointer. Cross-file functions are now declared once in xfs/xfs_mxfs_dlm.h so every caller is checked.
- mxfs_dir3_disk_has_extra_inum() answered 0 ("nothing of ours missing") for a block with more than 256 entries, the condition under which the subset guards drop a directory write; the index is sized from the block and an unknown answer is negative.
- Zoned filesystems are refused on every kernel: on 6.17/7.0 they mounted into a zone stub whose signature did not match its callers. Every stub in xfs_stubs.c now compiles against its real header (nine signatures corrected).
- Misleading-indentation fixes: the exclusive-holder trace was recorded twice on four inode-lock paths.
- 6.6 KB of name index moved off the stack in the two directory write merges (6.8 KB and 4.8 KB frames), and 1.5-2 KB arrays in four scans; 16 format specifiers fixed; lktdump rejects non-numbers; dead functions and variables removed.
- New scripts/extern_decl_audit.py checks every extern declaration against its definition.
- Clean 6.8 build: 236 warnings to 4 (1.0-1.3 KB frames). Platform builds (PVE 6.17/7.0, Alma/Rocky 9.8), usermode DLM tests 8/8, 2/tcp suite 29/30 with fio_perf's known budget margin (reruns pass at 105 s and 104 s).
- Defect queue: new D-WARN-SITES-PAST-LINE-65535 (xfs_mxfs_dlm.c exceeds the bug table's 16-bit line); fio_perf record updated with this run.

+++ xfs/libxfs/xfs_dir2.c
@@ -1992,11 +1992,11 @@ mxfs_dir_rename_fail(
 {
 	if (likely(!error) || !mp->m_mxfs_dlm)
 		return;
-	pr_warn_ratelimited("mxfs: P217-RENAME-FAILSITE site=%s rc=%d ino=%llu fmt=%u if_bytes=%u size=%lld nextents=%llu comm=%s — first failing helper inside xfs_dir_rename_children; the caller cancels a DIRTY transaction next\n",
+	pr_warn_ratelimited("mxfs: P217-RENAME-FAILSITE site=%s rc=%d ino=%llu fmt=%u if_bytes=%lld size=%lld nextents=%llu comm=%s — first failing helper inside xfs_dir_rename_children; the caller cancels a DIRTY transaction next\n",
 		site, error,
 		(unsigned long long)(subject ? subject->i_ino : 0),
 		subject ? subject->i_df.if_format : 0,
-		subject ? subject->i_df.if_bytes : 0,
+		subject ? (long long)subject->i_df.if_bytes : 0,
 		subject ? (long long)subject->i_disk_size : -1,
 		subject ? (unsigned long long)subject->i_df.if_nextents : 0,
 		current->comm);
